1. Restructed folders. 2. Added friend functions ( not tested ). 3.  Edited director.c to accomodate.
