+++ Server/C-remake/include/user_management.h
@@ -18,4 +18,8 @@ Error update_user(sqlite3 **db, char *userID, char *username, char *email, char 
 
 Error login(sqlite3 **db, char *username, char *password, String *output);
 
+Error add_friend(sqlite3 **db, char *user_id, char *friend_username);
+
+Error remove_friend(sqlite3 **db, char *user_id, char *friend_username);
+
 #endif //USER_MANAGEMENT_H
