+++ Server/C-remake/src/director.c
@@ -68,29 +68,55 @@ Error input_handler(sqlite3 **db, char* input, String *output) {
     }
     return status;
 
-  }else if(strcmp(inputToken, "test") == 0){
+  }else if(strcmp(inputToken, "addfriend") == 0) {
+    printf("%s\n", inputToken);
+    char *user_id = strtok(NULL, "|");
+    printf("User ID: %s\n", user_id);
+    char *friend_username = strtok(NULL, "|");
+    if(user_id == NULL || friend_username == NULL){
+      fprintf(stderr, "ERROR: Missing parameters for addfriend command.\n");
+      return (Error) {INVALID_ARGUMENT,
+                      "director.c/input_handler/ERROR: Missing parameters for addfriend command.\n"};
+    }
+    Error status = add_friend(db, user_id, friend_username);
+    if(status.code != OK){
+      fprintf(stderr, "ERROR: Failed to add friend.\n");
+      return status;
+    }
+    return status;
 
+  }else if (strcmp(inputToken, "removefriend") == 0){
+    printf("%s\n", inputToken);
+    char *user_id = strtok(NULL, "|");
+    char *friend_username = strtok(NULL, "|");
+    if(user_id == NULL || friend_username == NULL){
+      fprintf(stderr, "ERROR: Missing parameters for removefriend command.\n");
+      return (Error) {INVALID_ARGUMENT,
+                      "director.c/input_handler/ERROR: Missing parameters for removefriend command.\n"};
+    }
+    Error status = remove_friend(db, user_id, friend_username);
+    if(status.code != OK){
+      fprintf(stderr, "ERROR: Failed to remove friend.\n");
+      return status;
+    }
+    return status;
+
+  }else if(strcmp(inputToken, "test") == 0){
     Table_String table = {.data = malloc(64 * sizeof(String*)), .rows = 0, .cols = 0, .table_capacity = 64};
     Error state = describe_table(db, "events", &table);
-
-        printf("Testing\n");
-
+    printf("Testing\n");
     if(state.code != OK){
       free_table(&table);
       fprintf(stderr, "ERROR: Failed to describe table\n");
       return state;
     }
-
     state = convert_table_to_string(&table, output);
-
     if(state.code != OK){
       free_table(&table);
       fprintf(stderr, "ERROR: Failed to convert table to string\n");
       return state;
     }
-
     free_table(&table);
-
     return (Error){OK, "Success.\n"};
   }else{
     printf("Cause of ERROR --> %s | ERROR: Invalid command\n ", inputToken);
@@ -101,14 +127,11 @@ Error input_handler(sqlite3 **db, char* input, String *output) {
 
 Error convert_table_to_string(Table_String *table, String *result){
   char *state;
-
   for(int i = 0; i < table->rows; i++){
     for(int k = 0; k < table->cols; k++){
-
       if(result->capacity < result->length + strlen(table->data[i][k].data) + 2){
         result->capacity *= 2;
         char *new_data = realloc(result->data, result->capacity);
-
         if(new_data == NULL){
           fprintf(stderr, "ERROR: Failed to reallocate more memory for string data\n");
           return (Error) {MEMORY_ALLOCATION_ERROR, 
@@ -122,7 +145,6 @@ Error convert_table_to_string(Table_String *table, String *result){
         return (Error) {STRING_ERROR, 
                         "director.c/convert_table_to_string/ERROR: Failed to concatenate table data to resulting string.\n"};
       }
-      
       state = strncat(result->data, "|", 2);
       if(state == NULL){
         fprintf(stderr, "ERROR: failed to concatenate separating symbol to resulting string\n");
@@ -131,7 +153,6 @@ Error convert_table_to_string(Table_String *table, String *result){
       }
       result->length += (strlen(table->data[i][k].data) + 1);
     }
-
     state = strncat(result->data, "\n", 2);
     result->length += 1; // Increment length for the new line character
     if(state == NULL){
