+++ Server/Database/DB_Info.txt
@@ -1,0 +1,74 @@
+TIP: SHIFT LEFT-CLICK ON THE ARROWS ON THE SIDE TO COLLAPSE EVERYTHING AND THEN EXPAND THEM AS NEEDED FOR VIEWING PLEASURE
+
+id (int): The account number/When it was created
+
+username (String): The name of the user
+
+email (String): The email associated with the user
+
+password (String): The password obviously
+
+friends (String): The friends associated with the account
+    - Setup
+        - {string username}
+    - Example
+        - {Jeremy1223}, {LuisIsTheDatabaseArchitect}
+
+event (CSV String): The events planned
+    - Setup
+        - {string title, string date, string startTime, string endTime, string location, string publicityType, string[] invitees, string details, int id}
+            - date format
+                - "MM/DD/YYYY"
+            - startTime format
+                - "HH:MM:SS"
+            - endTime format
+                - "HH:MM:SS"
+            - location
+                - Format
+                    - "Long,Lat"
+                    - "12.24534241,-35.1243533241"
+                - Extra info
+                    - Lat: N
+                    - Long: W
+            - publicityType
+                - Types
+                    - FriendsOnly
+                        - All friends can see these events posted by you
+                    - Public
+                        - Anyone can see these
+                    - Private
+                        - Nobody except you and invitees can see this
+                - Format
+                    - "Type"
+                - Extra info
+                    - Public not added for demo
+                    - Public will only show to people who are within 20 miles(?)
+            - invitees
+                - Format
+                    - "Friend, Friend..."
+                - Extra info
+                    - invitees will automatically update depending on the publicityType, but also can have people manually invited or removed. The user set invitees takes full priority.
+            - id
+                - Format
+                    - num
+                - Extra info
+                    - literally just based off of n+1
+    - Example
+        - {"Party", "08/24/2025", "00:00:00", "24:00:00", "FriendsOnly", "["Friend1", Friend2", "Friend3"]", "Backflipping off of the edge of a pool so that the nuclear bombs don't hit me", 1},
+          {"Event 2", "02/13/2440", "23:10:56", "23:20:00", "Private", "N/A", "", 2}
+    - PROGRAMMER LOGIC
+        - Once the publicityType is selected the program should automatically select the invitees
+        - id should be automatically set based on the id of the last previously made event
+
+shared (string): Events shared with the user
+    - Setup
+        - {string inviter, int id}
+            - inviter
+                - The person who invited you to this event.
+            - id
+                - The id of the event in the inviters register  
+    - PROGRAMMER LOGIC
+        - Make the information appear on the bulletin board and in the users calendar.
+        - Check every time the page is loaded to see if the event still exists
+            - How to Check: Search the Username and event ID number and if event returns null then it doesn't exist
+
