+++ Server/Frontend/About/About.html
@@ -1,0 +1,3 @@
+<div>
+	<p>To be Implemented at a later date</p>
+</div>
