+++ Server/Frontend/Main/Main.css
@@ -1,0 +1,73 @@
+.login {
+    margin-top: 150;
+    margin-left: 60;
+    width: auto;
+    height: 750px;
+    background-image: url("../Images/login.png");
+    background-size: contain;
+    background-repeat: no-repeat;
+}
+
+.signup {
+    margin-top: 50;
+    margin-left: 40;
+    width: auto;
+    height: 750px;
+    background-image: url("../Images/signup.png");
+    background-size: contain;
+    background-repeat: no-repeat;
+}
+
+.bluepin {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+    width: 1%;
+}
+
+input.loginInput {
+    background: transparent;
+    border: none;
+    line-height: 2.65rem;
+    width: 49.45%;
+}
+
+div.loginInput {
+    margin-top: 310px;
+    margin-left: 9.5%;
+    margin-right: 9.5%;
+    width: auto;
+}
+
+button.loginButton {
+    display: block;
+    margin: auto;
+    padding: 1px 2%;
+}
+
+div.loginButton {
+    margin-top: 48%;
+    margin-bottom: 0;
+    position: relative;
+    text-align: center;
+}
+
+.signupButton {
+    padding: 66px 4.75555555%;
+    background: transparent;
+}
+
+.buttons {
+   font-size: 0px;
+   margin-top: 530px;
+   float: left;
+   width: 100%;
+}
+
+#gap1 {
+    margin-left: 10%;
+}
+
+#gap2 {
+    margin-left: 20%;
+}
