--- Website/About/About.html
@@ -1,3 +1,0 @@
-<div>
-	<p>To be Implemented at a later date</p>
-</div>
