--- Website/Home/Home.css
@@ -1,168 +1,0 @@
-.calendar {
-    width: 64.55%;
-}
-
-.monthdisplay {
-    position: relative;
-    display: inline-block;
-    background-color: white;
-    border: 5px solid;
-    width: 30%;
-    height: 10%;
-    margin-bottom: 0.4%;
-}
-
-.eventviewer {
-    border: 5px solid;
-    background-color: white;
-    width: 30%;
-    height: 70%;
-    float: right;
-    overflow-y: auto;
-}
-
-#eventviewertitle {
-    font-size: 36px;
-}
-
-table {
-    width: 100%;
-    height: 80%;
-    border-collapse: collapse;
-}
-
-tr {
-    width: 100%;
-    border: 5px solid;
-}
-
-td {
-    width: 14%;
-    height: auto;
-    background-color: white;
-    border: 5px solid;
-}
-
-th {
-    height: 1rem;
-    background-color: lightgrey;
-}
-
-.navbutton {
-    display: inline-block;
-    border: 0;
-    background-color: transparent;
-    height: 100%;
-    width: 10%;
-    font-size: 3rem;
-}
-
-#rightbutton {
-    position: absolute;
-    right: 0;
-}
-
-ul {
-    list-style-type: none;
-    margin-left: 0;
-    padding: 0;
-}
-
-.datenumber {
-    width: 100%;
-    height: 142px;
-    border: 0;
-    position: relative;
-}
-
-.calendarbox {
-    font-size: 7rem;
-    height: 100%;
-    width: 100%;
-}
-
-.eventheader {
-    background-color: #D6F6FF;
-    border: 1px solid;
-}
-
-.dropdownheader {
-    background-color: beige;
-    border: 1px solid;
-    width: 100%;
-    margin: 0;
-    padding: 0;
-}
-
-.dropdowncontent {
-    margin: 0;
-    padding: 0;
-    border: 1px solid;
-    background-color: lightgrey;
-}
-
-.dropdownbutton {
-    background-color: transparent;
-    border: none;
-    margin-left: auto;
-    margin-right: 0;
-}
-
-.eventpeople {
-    margin: 0;
-    padding: 0;
-}
-
-.eventdescription {
-    margin: 0;
-    padding: 0;
-}
-
-.nospacing {
-    margin: 0;
-    padding: 0;
-}
-
-.dropbtn {
-  background-color: #4CAF50;
-  color: white;
-  padding: 16px;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-}
-
-.menudropdown {
-  position: relative;
-  display: inline-block;
-}
-
-.menudropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 160px;
-  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-  right: 0;
-  z-index: 1;
-  align-items: center;
-}
-
-.menudropdown-content a {
-  color: black;
-  padding: 12px 16px;
-  text-decoration: none;
-  display: block;
-}
-
-.menudropdown-content a:hover {
-    background-color: #f1f1f1
-}
-
-.menudropdown:hover .menudropdown-content {
-  display: block;
-}
-
-.menudropdown:hover .dropbtn {
-  background-color: #3e8e41;
-}
